requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
python-dotenv==0.19.0
PyJWT==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
